requirements.txt
PyQt5
pyqrcode
Pillow